refactor(runtime): extract install module from plugin_manager
The install / removal / registry-listing cluster moves into
`plugin_manager/install.rs`:

- `install_plugin` (pub) — source-of-install dispatcher.
- `install_from_local`, `install_from_registry`, `install_from_git`
  — the three install paths.
- `validate_github_repo` — owner/repo allowlist + length sanity.
- `RegistryPluginEntry` struct + `list_registry_plugins` (pub async)
  + the `parse_plugin_i18n_blocks` parser + `fetch_registry_plugin_meta`
  + `registry_listing_cache_path` — registry-listing cache layer.
- `remove_plugin` (pub) — directory-tree delete with safe-path check.
- `install_requirements` (pub async) — Python `pip install -r` runner.
- `GitHubContent` struct + `download_github_entry` — recursive content
  walker used by the git / registry-via-github paths.

The five `pub` items are re-exported from plugin_manager.rs via
`pub use self::install::{...}` so the kernel and api routes keep
resolving `librefang_runtime::plugin_manager::install_plugin` /
`remove_plugin` / `install_requirements` /
`list_registry_plugins` / `RegistryPluginEntry`.

install.rs imports its registry deps via
`use super::registry::{default_registry_cache_ttl_secs,
fetch_verified_index, load_registry_cache, registry_cache_path,
save_registry_cache}` and the `HashMap` / `PluginI18n` types via
explicit `use` (these were inherited via `use super::*` before the
split but `super::*` from a child no longer re-exposes the parent's
imported `use` aliases).

The stranded `/// Install a plugin from a source.` doc-line that the
deletion range left at the boundary above `sha256_hex` is removed in
the same commit (fixes the `clippy::empty-line-after-doc-comments`
lint). Also removes the now-dead helper-block comment that doc-
referenced the moved `check_hooks_exist`.

plugin_manager.rs: 2,537 -> 1,738 LOC (-799). Now under the 2,000
LOC cap from #3710's acceptance criteria.

cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
